Moved grammar files to separate dir
